Allow client to send metrics to IPV6 addresses
